refactor(sandbox-fc): split netns pool modules
